pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyreadline3>=3.5.4",
    "readline>=6.2.4.2",
    "requests>=2.32.3",
    "selenium>=4.29.0",
]
